+++ tests/test_instrumentation/test_trace_back.py
@@ -10,7 +10,7 @@ from satella.instrumentation import Traceback
 
 class TestTraceback(unittest.TestCase):
     def test_no_exc(self):
-        tb = Traceback()
+        self.assertRaises(ValueError, Traceback)
 
     def test_tb(self):